Address review: test real behavior, not mock calls; cover the upload path
- refresh_access_token's tests now use a real temp .env file (monkeypatched
  ENV_PATH) instead of mocking set_key, and assert on the file's actual
  content — the mock-based version only verified set_key was *called* with
  plausible-looking arguments, not that .env was actually updated (or left
  alone) correctly.
- Add two tests exercising main()'s real upload path in-process (Strava
  mocked via `responses`, a real temp .env, monkeypatched argv) — the only
  path previously under test was --dry-run, so refresh -> upload -> poll ->
  mute being wired together correctly, the printed activity URL, and the
  --public flag's effect on muting were all completely untested.
- Fix pyright errors in test_strava_client.py surfaced while touching this
  file (responses' `.body` is typed BodyType | None, so `in`/`.split()`
  needed an isinstance narrowing first) and widen poll_upload's timeout_s
  to float, matching how it's actually used (compared against time.time())
  and tested (0.05s).

+++ sync_to_strava.py
@@ -243,7 +243,7 @@ def upload_activity(access_token: str, workout: dict, sport_type: str) -> dict:
     return resp.json()
 
 
-def poll_upload(access_token: str, upload_id: int, timeout_s: int = 30) -> dict:
+def poll_upload(access_token: str, upload_id: int, timeout_s: float = 30) -> dict:
     # Starts unset rather than only being assigned inside the loop: with a
     # very small timeout_s the loop body can run zero times, and the
     # TimeoutError below used to reference this while still unbound,
